add action/envelope.go&&envelope_test.go action/sealedenvelope.go&&sealedenvelope_test.go
